% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColorsMatrix.R
\name{addColorsMatrix}
\alias{addColorsMatrix}
\title{Annotation of a matrix with colors for contrast values}
\usage{
addColorsMatrix(resultsContrast, selectedComparison)
}
\arguments{
\item{resultsContrast}{A matrix of all the values that correspond to the relative presence of RNA for each transcript analyzed}

\item{selectedComparison}{The selected comparison to display}
}
\value{
The object resultsContrast annotated with an associated color for the selected comparison
}
\description{
Annotation of a matrix with colors for contrast values
}
\examples{
resultsContrast <-
selectedComparison <-
resultsContrastRamp <- addColorsMatrix(resultsContrast,selectedComparison)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[rnaseqApp]{colorNumericValues}}
}
\keyword{colors}

