% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColorsMatrixResults.R
\name{addColorsMatrixResults}
\alias{addColorsMatrixResults}
\title{Annotation of a matrix with colors from standardized values}
\usage{
addColorsMatrixResults(results, selectedComparison)
}
\arguments{
\item{results}{A matrix of all the values that correspond to the presence of RNA for each transcript analyzed}

\item{selectedComparison}{The selected comparison to display}
}
\value{
The object resultsContrast annotated with an associated color for the selected comparison
}
\description{
Annotation of a matrix with colors from standardized values
}
\examples{
results <- expr.matrix <- readRDS("data/results_LGVD.rds")
selectedComparison <-
resultsContrastRamp <- addColorsMatrix(resultsContrast,selectedComparison)
}
\author{
Simon J Pelletier
}
\keyword{colors}

