% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy2ensembl.R
\name{affy2ensembl}
\alias{affy2ensembl}
\title{Convert affyId to ensemblID}
\usage{
affy2ensembl(affyID)
}
\arguments{
\item{affyId}{Vector of all affyId to be converted to ensemblID}
}
\value{
The object resultsContrast annotated with an associated color for the selected comparison
}
\description{
Convert affyId to ensemblID
}
\author{
Simon J Pelletier
}
\keyword{convert}

