% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ensembl.R
\name{annotate_ensembl}
\alias{annotate_ensembl}
\title{Takes an ensembl_gene_id}
\usage{
annotate_ensembl(IDs, type)
}
\arguments{
\item{IDs}{All IDs (rat or human); all are ensembl_gene_id}
}
\value{
The object resultsContrast annotated with an associated color for the selected comparison
}
\description{
Takes an ensembl_gene_id
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
IDs <- rownames(expr.matrix)
annotation1 <- annotate_ensembl(IDs)
annotations <- annotation1[[1]]
go <- annotation1[[2]]
genes_annotation_unique <- annotation1[[3]]
typeID <- annotation1[[4]]

}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{colorNumericValues}}
}
\keyword{IDs}
\keyword{ensembl}

