% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{annotation_biomart}
\alias{annotation_biomart}
\title{Biomart annotation}
\usage{
annotation_biomart(IDs, specie, attribute, mart)
}
\description{
Biomart annotation
}
\examples{
specieEnsembl <- "hsapiens_gene_ensembl"
attribute <- "affy_hg_u95av2"
gset <- getGEO("GSE12654")
exprset <- gset[[1]]
type <- "ensembl_gene_id"
expr.matrix <- exprs(exprset)
IDs <- rownames(expr.matrix)[1:10]
x<-annotation_biomart(IDs,specieEnsembl,attribute)
}
\author{
Simon J Pelletier
}

