% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_geo.R
\name{annotation_geo}
\alias{annotation_geo}
\title{Get the annotations used in a NCBI GEO dataset}
\usage{
annotation_geo(geoAccession)
}
\arguments{
\item{geoAccession}{The platform ID}
}
\value{
A table with all the annotation for a platform
}
\description{
Get the annotations used in a NCBI GEO dataset
}
\examples{

x=annotation_geo('GPL10558') #Illumina HumanHT-12 V4.0 expression beadchip
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[GEOquery]{getGEO}} ,
\code{\link[GEOquery]{Meta}} ,
\code{\link[GEOquery]{Table}} ,
\code{\link[annotate]{readGEOAnn}}
}
\keyword{annotation}
\keyword{geo}

