% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringSummary.R
\name{clusteringSummary}
\alias{clusteringSummary}
\title{Get the annotations used in a NCBI GEO dataset}
\usage{
clusteringSummary(expr.matrix, pval, names)
}
\arguments{
\item{expr.matrix}{A matrix of numeric values. Rows are genes, columns are samples}

\item{pval}{Alpha value (selected p-value)}
}
\value{
A table with all the annotation for a platform
}
\description{
Get the annotations used in a NCBI GEO dataset
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
pval <- 0.05
summaryClust <- clusteringSummary(expr.matrix,pval)

gset <- getGEO("GSE44593", GSEMatrix =TRUE)
exprset <- gset[[1]]
expr.matrix <- exprs(exprset)
names1 <- get_names(sampleNames(exprset))
groupBy <- "group"
noCluster <- 3
fileContent <- "Expression Set"
names1 <- get_names(expr.matrix)
selectedVariables <- NULL
comparisonsTable <- comparisonsPheno(exprset)[[2]]
namesTable <- comparisonsTable
names2 <- namesSelectedComparisons(namesTable)
summaryClust <- clusteringSummary(expr.matrix,pval,names2)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{fuzzy_clustering}} ,
\code{\link[stats]{phyper}}
}
\keyword{Hypergeometric}
\keyword{cluster}
\keyword{clustering}
\keyword{fuzzy}

