% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_groups.R
\name{color_groups}
\alias{color_groups}
\title{Add color to each sample. Samples with the same color belong to the same group}
\usage{
color_groups(names)
}
\arguments{
\item{names}{Names of each sample}

\item{names.unique}{Name of each group}
}
\value{
A vector of strings corresponding to the color of all samples.
}
\description{
Add color to each sample. Samples with the same color belong to the same group
}
\author{
Simon J Pelletier
}

