% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorNumericValues.R
\name{colorNumericValues}
\alias{colorNumericValues}
\title{Transform intensity values in a gradient of color from green to red.}
\usage{
colorNumericValues(values)
}
\arguments{
\item{values}{Intensity values to be transform in color gradients}
}
\value{
A vector of colors. The darkest green values are the lowest and the darkest red values are the highest.
}
\description{
Transform intensity values in a gradient of color from green to red.
}
\author{
Simon J Pelletier
}
\keyword{comparisons}

