% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonsPheno.R
\name{comparisonsPheno}
\alias{comparisonsPheno}
\title{Transform intensity values in a gradient of color from green to red.}
\usage{
comparisonsPheno(exprset)
}
\arguments{
\item{exprset}{Expression set containing all the information on a dataset}
}
\value{
A vector of colors. The darkest green values are the lowest and the darkest red values are the highest.
}
\description{
Transform intensity values in a gradient of color from green to red.
}
\examples{
gset <- getGEO('GSE54839', GSEMatrix =TRUE) #GSE61276 GSE12654
exprset <- gset[[1]]
comparisons <- comparisonsPheno(exprset)[[1]]
comparisonsTable <- comparisonsPheno(exprset)[[2]]
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[GEOquery]{getGEO}}
\code{\link[Biobase]{ExpressionSet}}
}
\keyword{comparisons}

