% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionID.R, R/conversionTable.R
\name{conversionID}
\alias{conversionID}
\title{Convert any type of ID to ensembl ids}
\usage{
conversionID(specie_genome = "rnorvegicus_gene_ensembl",
  attribute = "efg_agilent_wholegenome_4x44k_v1")

conversionID(specie_genome = "rnorvegicus_gene_ensembl",
  attribute = "efg_agilent_wholegenome_4x44k_v1")
}
\arguments{
\item{specie_genome}{The ensembl ID of the genome studied e.g. rnorvegicus_gene_ensembl}

\item{attribute}{The type of ID to convert to ensembl_gene_id and ensembl_transcript_id}

\item{specie_genome}{The ensembl ID of the genome studied e.g. rnorvegicus_gene_ensembl}

\item{attribute}{The type of ID to convert to ensembl_gene_id and ensembl_transcript_id}
}
\value{
A 3 column data.frame: ensembl_gene_id ensembl_transcript_id attribute(id chosen)

A 3 column data.frame: ensembl_gene_id ensembl_transcript_id attribute(id chosen)
}
\description{
Convert any type of ID to ensembl ids

Convert any type of ID to ensembl ids
}
\examples{
specie_genome = "hsapiens_gene_ensembl"
attribute <- "affy_hg_u95av2"
externalSymbol <- "hgnc_symbol"
ensemblID <- conversionID(specie_genome,attribute,externalSymbol)
specie_genome = "hsapiens_gene_ensembl"
attribute <- "affy_hg_u95av2"
ensemblID <- conversionTable(specie_genome,attribute)
}
\author{
Simon J Pelletier

Simon J Pelletier
}

