% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_contrasts.R
\name{design_contrasts}
\alias{design_contrasts}
\title{Create a design matrix for linear models analysis}
\usage{
design_contrasts(expr.matrix)
}
\arguments{
\item{names}{A vector of the names of all samples}
}
\value{
An experimental design matrix
}
\description{
Create a design matrix for linear models analysis
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
names <- get_names(colnames(expr.matrix))
design <- design_contrast(names)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{get_names}}
}
\keyword{design}
\keyword{limma}
\keyword{linear}

