% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2cytoscape.R
\name{export2cytoscape}
\alias{export2cytoscape}
\title{Export results in cytoscape format}
\usage{
export2cytoscape(expr.matrix, results, threshold = 0.9,
  type = "ensembl_gene_id")
}
\arguments{
\item{expr.matrix}{A matrix of numeric values. Rows are genes, columns are samples}

\item{results}{A topTable with only the genes considered significant (according to
the p-value and logFC limits provided in the function results_topTable)}

\item{threshold}{The lowest value of correlation to make the link
considered significantly high}

\item{type}{The type of ID used, e.g.}
}
\value{
The results in cytoscape format
}
\description{
Export results in cytoscape format
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
results <- readRDS("data/results_LGVD.rds")
export2cytoscape <- export2cytoscape(expr.matrix,results[[1]])
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[WGCNA]{adjacency}}
}
\keyword{cytoscape}
\keyword{export}

