% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprToBind.R
\name{exprToBind}
\alias{exprToBind}
\title{Binds an expression matrix with colors corresponding to the module of the elements
in the rows (genes,transcripts...)}
\usage{
exprToBind(bnet, expr.matrix)
}
\arguments{
\item{bnet}{A list of WGCNA results}

\item{expr.matrix}{A matrix of expression values. Rows are genes, columns are samples}

\item{typeID}{Type of IDs used as rows in the expression matrix.Default="ensembl_gene_id"}
}
\value{
An expression matrix additonnal informations with two columns added as the first columns
 of the matrix. The columns are, in order:
\describe{
 \item{module}{Contain the module of the element (gene, transcript...)}
 \item{ID}{contain the IDs used afterwards to merge all of the additional
           annotations in subsequent operation in the pipeline}
}
}
\description{
Binds an expression matrix with colors corresponding to the module of the elements
in the rows (genes,transcripts...)
}
\examples{
bnet <- readRDS("data/bnet_LGVD.rds")
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
typeID <- "ensembl_gene_id"
expr.toBind <- exprToBind(bnet,expr.matrix,typeID)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[limma]{topTable}}

\code{\link[ggplot2]{ggplot}}
}
\keyword{cytoscape}
\keyword{export}

