% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_clustering.R
\name{fuzzy_clustering}
\alias{fuzzy_clustering}
\title{Find the groups using fuzzy clustering ##FIND AUTOMATICALLY THE NUMBER OF CORES}
\usage{
fuzzy_clustering(expr.matrix, noCluster = 1)
}
\arguments{
\item{expr.matrix}{A matrix of expression values. Rows are genes, columns are samples}

\item{noCluster}{The number of clusters to be used. ##COULD BE DETERMINATED AUTOMATICALLY??}
}
\value{
The function return a list of 3 objects
\describe{
 \item{clusters}{Vector in which every sample is attributed to the most likely cluster}
 \item{in_mfuzz}{Expression set}
 \item{cl_mfuzz}{Details of fuzzy cmeans}
}
}
\description{
Find the groups using fuzzy clustering ##FIND AUTOMATICALLY THE NUMBER OF CORES
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
noCluster <- 3
fuzzy <- fuzzy_clustering(expr.matrix,noCluster)
fuzzy$clusters
fuzzy$in_mfuzz
fuzzy$cl_mfuzz

}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[Mfuzz]{standardise}} ,
\code{\link{get_names}} ,
\code{\link{e1071}}
}
\keyword{Hypergeometric}
\keyword{cluster}
\keyword{clustering}
\keyword{fuzzy}

