% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{List of informations on Gene Ontology of selected genes}
\usage{
geneOntology(resultsList, go, type = "ensembl_gene_id", ngenes)
}
\arguments{
\item{go}{An object that contain all gene ontology informations for every gene}

\item{ngenes}{A list of WGCNA results}

\item{results}{A topTable with only the genes considered significant (according to
the p-value and logFC limits provided in the function results_topTable)}

\item{typeID}{Type of IDs used as rows in the expression matrix.Default="ensembl_gene_id"}
}
\value{
A list of gene ontology informations for all comparisons determinated my the
experimental design. For example, if there are 3 groups (a,b,c), 3 comparisons are
possible (a-b,a-c,b-c). The object return is a list containing 3 lists named "a-b", etc.
Each of these lists contain the following objects:
\describe{
 \item{geneGO.selected}{For every gene ontology that imply at least 1 significant gene, }
 \item{GOarray.percent}{}
 \item{GOarray}{}
 \item{GOarray_CellComponent}{}
 \item{GOarray_MolecularFunction}{}
 \item{GOarray_BiologicalProcess}{}
}
}
\description{
List of informations on Gene Ontology of selected genes
}
\examples{
results <- readRDS("data/results_LGVD.rds")
go <- read.csv("annotations/go.csv",header=FALSE)
typeID <- "ensembl_gene_id"
colnames(go) = c(typeID,"go_term_name","go_domain")
ngenes <- 26689
ontologyInfos <- geneOntology(results,go,typeID,ngenes)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{annotate_ensembl}}
\code{\link{annotate_geo}}
\code{\link{results_topTable}}
\code{\link[limma]{topTable}}
}
\keyword{cytoscape}
\keyword{export}

