% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparisons.R
\name{get_comparisons}
\alias{get_comparisons}
\title{Take all the name and export all possible comparisons}
\usage{
get_comparisons(names)
}
\arguments{
\item{names}{Names of all groups compared}
}
\value{
A vector of all the possible comparisons in the format names1//name2 (VERIFY!!)
}
\description{
Take all the name and export all possible comparisons
}
\author{
Simon J Pelletier
}
\keyword{comparisons}

