% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get names of each sample and group names}
\usage{
get_names(x)
}
\arguments{
\item{x}{Matrix or vector
column name has to be separeted}
}
\value{
The name of each groups.
}
\description{
Get names of each sample and group names
}
\examples{
EXPRESSION MATRIX
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
get_names(expr.matrix)

EXPRESSION SET
gset <- getGEO("GSE12654", GSEMatrix =TRUE)
exprset <- gset[[1]]
names1 <- get_names(sampleNames(exprset))
}
\author{
Simon J Pelletier
}
\keyword{names}
\keyword{names.unique}

