% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2Contrast.R
\name{lm2Contrast}
\alias{lm2Contrast}
\title{Linear Model analysis with limma}
\usage{
lm2Contrast(expr.matrix, design)
}
\arguments{
\item{expr.matrix}{A matrix of standardized data. Columns = samples, rows = genes,transcripts,CpG...}

\item{design}{The design of the experiment (object)}
}
\value{
Linear model contrast ...
\describe{
 \item{lm2.contrast}{MArrayLM object(limma) result of the eBayes function.}
 \item{contrasts}{Vector of contrasts}
 \item{contrast.matrix}{Matrix of contrasts}
}
}
\description{
Linear Model analysis with limma
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
design <- design_contrasts(get_names(colnames(expr.matrix)))
results_list <- results_topTable()
lm2 <- lm2Contrast(expr.matrix,design)
lm2.contrast = lm2[[1]]
contrasts=lm2[[2]]
contrast.matrix=lm2[[3]]
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{design_contrasts}}
\code{\link{makeContrasts}}
\code{\link[limma]{lmFit}}
\code{\link[limma]{MArrayLM}}
}
\keyword{limma}
\keyword{linear}

