% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot.R
\name{MAPlot}
\alias{MAPlot}
\title{Generates a MAPlot with ggplot}
\usage{
MAPlot(topTable3, logfc, pval)
}
\arguments{
\item{topTable3}{topTable results for every gene}

\item{logfc}{log2 fold-change limit}

\item{pval}{p-value limit}
}
\value{
A list of WGCNA results
}
\description{
Generates a MAPlot with ggplot
}
\examples{
topTable3 <-
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[limma]{topTable}} ,
\code{\link[ggplot2]{ggplot}}

\code{\link[limma]{plotMA}}{ : Similar function in the limma package}
}
\keyword{cytoscape}
\keyword{export}

