% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_summary.R
\name{modules_summary}
\alias{modules_summary}
\title{Returns the summary for a module}
\usage{
modules_summary(resultTable, expr.toBind)
}
\arguments{
\item{expr.toBind}{Expression data.matrix with supplemental colums: module and ID}

\item{results}{The results of a comparison between two groups.}
}
\description{
Returns the summary for a module
}
\examples{
results <- readRDS("data/results_LGVD.rds")
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
bnet <- readRDS("data/bnet_LGVD.rds")

expr.toBind <- cbind(module=bnet$colors,ensembl_gene_id=rownames(expr.matrix),expr.matrix)
result <- results[[1]]
modulesTable <- modules_summary(results[[1]],expr.toBind)

expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
type = "ensembl_gene_id"
design <- design_contrasts(get_names(colnames(expr.matrix)))
lm2 <- lm2Contrast(expr.matrix,design)
lm2.contrast = lm2[[1]]
contrasts=lm2[[2]]
contrast.matrix=lm2[[3]]
names <- get_names(expr.matrix)[[1]]
specie = "rnorvegicus"
attribute <- "ensembl_gene_id"
externalSymbol <- "rgd_symbol"
annotationFile <- paste0("annotations/databases/",specie,"_ensembl_",attribute,".csv")
#ensemblTable <- annotation_biomart(rownames(expr.matrix),specie,attribute)
ensemblTable <- read.csv(annotationFile)
genes <- ensemblTable[as.character(ensemblTable[,attribute]) != "",]
annotation <- annotate_ensembl(genes[,1],type)
annotations <- annotation[[1]]
go <- annotation[[2]]
genes_annotation_unique <- annotation[[3]]

maskModule <- as.numeric(as.character(results[[1]]$module)) == 
resultsModule <- list(resultsModule=results[[1]][maskModule,])
ontologyModule <- geneOntology(resultsModule,go,type,length(rownames(voom.matrix)))
topModuleGO <- cbind(rownames(ontologyModule[[1]][[2]]),ontologyModule[[1]][[2]])
resultsModuleOntology <- resultsByOntology(1,resultsModule[[1]],ontologyModule[[1]])
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[stats]{phyper}}
}
\keyword{annotation}
\keyword{geo}

