% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesSelectedComparisons.R
\name{namesSelectedComparisons}
\alias{namesSelectedComparisons}
\title{Create a vector of the group membership for each sample}
\usage{
namesSelectedComparisons(namesTable)
}
\arguments{
\item{namesTable}{data.frame with a single column containing group names}
}
\value{
A vector of the group membership for each sample.The name of each
element is the initial name of the sample
}
\description{
Takes a data.frame with a single column containing strings, removes the whitespace
of all strings and returns it as a vector
}
\examples{
gset <- getGEO('GSE61276', GSEMatrix =TRUE) #GSE61276 GSE12654
exprset <- gset[[1]]
comparisons <- comparisonsPheno(exprset)[[1]]
comparisonsTable <- comparisonsPheno(exprset)[[2]]
namesTable <- comparisonsTable
names2 <- namesSelectedComparisons(namesTable)
}
\author{
Simon J Pelletier
}

