% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.graphite.R
\name{network.graphite}
\alias{network.graphite}
\title{Create a graphNEL graph with directed edges from the results for a specified reaction}
\usage{
network.graphite(pSymbol, expr.matrix, toptableNodes, typeID, threshold,
  selectedReaction, toptable3, selectedComparison, expr.toBind, pvalue)
}
\arguments{
\item{pSymbol}{Symbol IDs}

\item{expr.matrix}{Expression set containing all the information on a dataset}

\item{toptableNodes}{topTable with added column of the reaction nodes it belongs to}

\item{typeID}{Type of IDs used as rows in the expression matrix.Default="ensembl_gene_id"}

\item{threshold}{Threshold of coexpression}

\item{selectedReaction}{Selected reaction from the topReactions}

\item{selectedComparison}{Comparison studied. Name of the matrices (topTables) in the list "results"}

\item{expr.toBind}{Expression data.matrix with supplemental colums: module and ID}

\item{pvalue}{p-value limit}

\item{topTable3}{topTable results for every gene}
}
\value{
A graphNEL graph with directed edges
}
\description{
Create a graphNEL graph with directed edges from the results for a specified reaction
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
topReactions <- readRDS("data/reactions_mfuzz.rds")
topTable3 <- readRDS("data/topTable3_LGVD.rds")
selectedReaction <- "RHO GTPases activate PAKs"
ratReactome <- pathways("rnorvegicus", "reactome")
p <- ratReactome[[selectedReaction]]
pSymbol <- convertIdentifiers(p, "SYMBOL")
toptableNodes = cbind(
   topTable3[[1]][match(nodes(pSymbol),topTable3[[1]]$symbol),],
   Reaction = rep(selectedReaction[[1]],length(nodes(pSymbol)))
)
toptableNodes1 = toptableNodes[!is.na(toptableNodes[,1]),]
typeID = "ensembl_gene_id"
threshold = 0.8

network.graphite(pSymbol,expr.matrix,toptableNodes,typeID,threshold,selectedReaction,topTable3,1,expr.toBind,pvalue)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{export2cytoscape}}
\code{\link{graphite}}
}
\keyword{graphite}
\keyword{network}

