% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberNames.R
\name{numberNames}
\alias{numberNames}
\title{Add increasing number at the end of all strings. Duplicated names are uniquely named by incrementing
the number of each new duplicate found.}
\usage{
numberNames(names)
}
\arguments{
\item{names}{Strings to add numbers}
}
\value{
Vector of the numbered strings
}
\description{
Add increasing number at the end of all strings. Duplicated names are uniquely named by incrementing
the number of each new duplicate found.
}
\author{
Simon J Pelletier
}

