% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_list.R
\name{order_list}
\alias{order_list}
\title{Order the element of a list by size in decreasing order}
\usage{
order_list(list)
}
\arguments{
\item{list}{List to order alphabetically}
}
\value{
The list ordered
}
\description{
Order the element of a list by size in decreasing order
}
\examples{
list <- list(a="toto",b=c("tyro","totor","cipaille","maple_syrup"),c=c("coconut","rdl"))
order_list(list)
}

