% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Displays a Princpial component analysis graph. USE GGPLOT!!}
\usage{
PCA(expr.matrix, names, colorGroups, namesColor, cex.size)
}
\arguments{
\item{expr.matrix}{Expression set containing all the information on a dataset}

\item{colorGroups}{Color associated with each sample}
}
\description{
Displays a Princpial component analysis graph. USE GGPLOT!!
}
\seealso{
\code{\link[vegan]{rda}}
\code{\link[vegan]{ggplot}}
}

