% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrectAttribution.R
\name{plotCorrectAttribution}
\alias{plotCorrectAttribution}
\title{Plot the groups correctly attributed}
\usage{
plotCorrectAttribution(fuzzy)
}
\description{
Plot the groups correctly attributed
}

