% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrectSubsample.R
\name{plotCorrectSubsample}
\alias{plotCorrectSubsample}
\title{Plot the subsamples correctly attributed}
\usage{
plotCorrectSubsample(fuzzy)
}
\description{
Plot the subsamples correctly attributed
}

