% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibleComparisons.R
\name{possibleComparisons}
\alias{possibleComparisons}
\title{Possible Comparisons}
\usage{
possibleComparisons(names)
}
\arguments{
\item{names}{A vector of the names of all samples}
}
\value{
A vector of all possible comparisons
}
\description{
Find all the possible pairs of groups to compare
}
\author{
Simon J Pelletier
}
\keyword{comparisons}

