% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA_factors.R
\name{RDA_factors}
\alias{RDA_factors}
\title{RDA plots
Creates a biplot}
\usage{
RDA_factors(expr.matrix, names)
}
\arguments{
\item{expr.matrix}{A matrix of standardized data. Columns = samples, rows = genes,transcripts,CpG...}
}
\description{
RDA plots
Creates a biplot
}
\examples{
expr.matrix <- readRDS('data/expr_matrix_LGVD.rds')
#' fileContent <- "Expression Matrix"
RDA_factors(expr.matrix[1:100,])
}

