% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactions.R
\name{reactions}
\alias{reactions}
\title{Top reactions}
\usage{
reactions(results, entrezgene_ids)
}
\arguments{
\item{results}{Results list (topTables)}

\item{entrezgene_ids}{All ids, entrez format}
}
\description{
Finds top reactions of each results table
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
annotation1 = annotate_ensembl(rownames(expr.matrix))
annotations=annotation1[[1]]
results <- readRDS("data/results_LGVD.rds")
entrezgene_ids <- unique(annotations$entrezgene_id)
reactome <- pathways("rnorvegicus", "reactome")
prepareSPIA(reactome, "reactome",print.names=TRUE)
res <- reactions(results,entrezgene_ids)
}
\seealso{
\code{\link[graphite]{pathways}}
\code{\link[graphite]{prepareSPIA}}
}

