% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroupingSamples.R
\name{regroupingSamples}
\alias{regroupingSamples}
\title{Regroup}
\usage{
regroupingSamples(fileContent, groupBy, expr.matrix, selectedVariables,
  noCluster = NULL, names1, comparisonsTable = NULL, exprset = NULL)
}
\description{
Regroup
}
\examples{
gset <- getGEO("GSE54839", GSEMatrix =TRUE,getGPL = FALSE) #GSE61276 GSE12654 GSE54839 Methylation: GSE18111

exprset <- gset[[1]]
expr.matrix <- exprs(exprset)
names1 <- get_names(sampleNames(exprset))
groupBy <- "group"
#expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
noCluster <- 3
fileContent <- "Expression Set"
names1 <- get_names(expr.matrix)
selectedVariables <- "characteristics_ch1"
selectedReplicates <- "source_name_ch1"
comparisonsTable <- comparisonsPheno(exprset)[[2]]
expr.matrix2 <- unduplicate(comparisonsTable,selectedReplicates,expr.matrix)
namesTable <- comparisonsTable
names2 <- namesSelectedComparisons(namesTable)
}

