% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsByOntology.R
\name{resultsByOntology}
\alias{resultsByOntology}
\title{Results by ontology
Return the rows from the results topTable}
\usage{
resultsByOntology(selectedOntology, result, ontology)
}
\arguments{
\item{selectedOntology}{The ontology term selected}

\item{ontology}{An object}

\item{results}{The results of one comparison}
}
\description{
Results by ontology
Return the rows from the results topTable
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
results <- readRDS("data/results_LGVD.rds")
annotation1 = annotate_ensembl(rownames(expr.matrix))
go=annotation1[[2]]
typeID=annotation1[[4]]
result <- results[[1]]
ontology <- geneOntology(results,go,typeID,nrow(expr.matrix))
selectedOntology <- names(ontology[[1]][[3]])[1]
resultsByOntology <- resultsByOntology(selectedOntology,result,ontology[[1]])
}
\seealso{
\code{\link{}}
}

