% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTopReaction.R
\name{selectTopReaction}
\alias{selectTopReaction}
\title{Select a reaction from a top reactions table}
\usage{
selectTopReaction(selectedComparison, topReactions)
}
\arguments{
\item{selectedComparison}{Selected Comparison}

\item{topReactions}{Top reactions}
}
\value{
The selected top reactions informations
}
\description{
Select a reaction from a top reactions table
}
\examples{
selectedComparison <- "a2-a1"
topReactions <- readRDS("data/reactions_mfuzz.rds")
selectedTopReactions <- selectTopReaction(selectedComparison,topReactions)
head(selectedTopReactions)
}
\author{
Simon J Pelletier
}

