% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifRows.R
\name{signifRows}
\alias{signifRows}
\title{Select in an expression matrix only the rows with significant elements (genes,transcripts...)}
\usage{
signifRows(expr.matrix, selectedElements)
}
\arguments{
\item{expr.matrix}{A matrix of standardized data. Columns = samples, rows = genes,transcripts,CpG...}

\item{resultsSummary}{A matrix of data from every gene that is significanlty different}
}
\value{
The significant rows of the initial expression matrix
between at least two groups. Columns = samples, rows = genes,transcripts,CpG...
}
\description{
Select in an expression matrix only the rows with significant elements (genes,transcripts...)
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
results <- readRDS("data/results_LGVD.rds")
topTable3 <- readRDS("data/topTable3_LGVD.rds")
resultsSummary <- results_summary(results,topTable3)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[limma]{topTable}}
\code{\link{results_summary}}
}

