% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unduplicate.R
\name{unduplicate}
\alias{unduplicate}
\title{Create a new expression matrix with means of technical replicates}
\usage{
unduplicate(comparisonsTable, selectedReplicates, expr.matrix)
}
\description{
Create a new expression matrix with means of technical replicates
}
\examples{
gset <- getGEO('GSE54839', GSEMatrix =TRUE) #GSE61276 GSE12654
exprset <- gset[[1]]
expr.matrix <- exprs(exprset)
comparisons <- comparisonsPheno(exprset,NULL)[[1]]
comparisonsTable <- comparisonsPheno(exprset,NULL)[[2]]
selectedReplicates <- "source_name_ch1"
}

