% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot.2}
\alias{volcanoPlot.2}
\title{Generates a volcanoplot with ggplot}
\usage{
volcanoPlot.2(topTable3, logfc = 1, pval = 0.05)
}
\arguments{
\item{topTable3}{topTable results for every gene}

\item{logfc}{log2 fold-change limit}

\item{pval}{p-value limit}
}
\value{
A list of WGCNA results
}
\description{
Generates a volcanoplot with ggplot
}
\examples{
topTable3 <- readRDS("data/topTable3.rds")
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link[limma]{topTable}} ,
\code{\link[ggplot2]{ggplot}}

\code{\link[limma]{plotMA}}{ : Similar function in the limma package}
}
\keyword{cytoscape}
\keyword{export}

