% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wes.R
\name{wes}
\alias{wes}
\title{Simple function to get color palettes from the wesanderson package,
which contain palettes from wes anderson movies}
\usage{
wes()
}
\description{
Simple function to get color palettes from the wesanderson package,
which contain palettes from wes anderson movies
}

