% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_modules.R
\name{WGCNA_modules}
\alias{WGCNA_modules}
\title{Find to which genes belong to the same modules}
\usage{
WGCNA_modules(expr.matrix, pow, names, deepSplit = 2, minModSize = 30,
  saveTOMs = TRUE, minKMEtoStay = 0, mergeCutHeight = 0.25,
  numericLabels = TRUE, pamRespectsDendro = FALSE)
}
\arguments{
\item{expr.matrix}{A matrix of expression values. Rows are genes, columns are samples}

\item{ncores}{Number of cores available for parallel programming (foreach function)}

\item{names.unique}{Names of all groups of sample}
}
\value{
A list of WGCNA results
}
\description{
Find to which genes belong to the same modules
}
\examples{
expr.matrix <- readRDS("data/expr_matrix_LGVD.rds")
results <- readRDS("data/results_LGVD.rds")

bnet = WGCNA_modules(expr.matrix,pow=3,names=names2)
}
\author{
Simon J Pelletier
}
\seealso{
\code{\link{WGCNA}}
}
\keyword{cytoscape}
\keyword{export}

